Fix bug in which [source foo.l] was interpreting foo.l as Tcl code, not L code.
The problem was that there are now two paths in which [source] is implemented, the old one
and the new NRE one. Due to backward compatibility, both need to have this code.


When in Tclland, do like the Tcl-ers. I.e. copy and paste the code in FSEvalFileEx() to 
the new NRE function so that we have the .l magic in both codepaths.

bk: 49b81bc2dobwYbnS_fp1w8t4qI1F3w