L: possible fix for L test failures seen on freebsd7 (and perhaps other freebsd's) caused by environment variables getting the wrong values.
This isn't caused by L and is reproducible with pure Tcl code.  My
thinking is that it's caused by us calling putenv() in tclPipe.c for
bk-specific reasons related to encrypted scripts.  This doesn't play
well with Tcl bypassing getenv/putenv but instead manipulating the
environ array directly (the first thing it does is to copy it and
point environ to the copy).  I've read that the Posix standard says
that if an application modifies environ, then the behavior of getenv()
is undefined, and we may be hitting this case.

The work-around is for us to simply call Tcl_SetEnv() instead of
setenv(), which modifies environ directly and avoids mixing and
matching setenv and writing to environ directly.

Here's a Tcl test case to illustrate the bug.  Put this in the file
printenv.tcl:

catch {puts $::env(L_TRACE)}

And then run this (thanks to Damon for the bells and whistles):

cd [file dirname [info script]]
puts "CWD =[pwd]"

set tclsh [info nameofexecutable]
puts "TCLSH = $tclsh"

puts [exec $tclsh printenv.tcl 2>@stdout]
set ::env(L_TRACE) fntrace10_1.out
puts [exec $tclsh printenv.tcl 2>@stdout]
puts $::env(L_TRACE)

On freebsd7 I saw it printing "fntrace10_1.outfntrace10_1.out" instead
of "fntrace10_1.out", and on some runs there was additional garbage.
This was causing the L fntrace-10.1 test to fail and the fntrace-10.2
test to hang.

bk: 4f7daf05_wMaIjVFfqmqa25ldMNVog