Add a real diff-engine to deal with subline highlighting in the GUI tools. Note that you can't do this in C because otherwise it wouldn't handle unicode properly (Tcl does).
The algorithm is based on the paper "An O(NP) Sequence 
Comparison Algorithm" by Sun Wu, Udi Manber, and Gene Myers,
and it was chosen because even thou it requires O(n^2) space
its running time is O(np) and is much simpler to implement
than Myers (what ndiff uses). Since we're dealing with lines
it's okay.

The rest is a bunch of heuristics to clean up the diffs, align
them and make them more pleasant. There is some special handling
of whitespace (to differentiate tabs vs 8 spaces in whitespace
only diffs). You need to see it to get it (run i.difftool-sl).

bk: 51817ee7vrou23bmBP2oI2Fg39RQaQ