L: deep-dive bytecode refactoring for both better performance and code clarity.
Abstract out details of the L deep-ptr type and create two
implementations, one for a fast path and one slower.  Also do far less
tricky code with the ref counts; this complexity was making the code
very hard to understand and maintain, and doesn't give a useful perf
win anyway.  Streamline the L sizes stack which was the next slowest
thing.  For an array or string index expression, don't generate
bytecodes for pushing/popping the array or string size if the
expression does not use END; this trims about another 30% off the time
to index into an array when END is not used.

While in here, do another attempted fix for a deep-dive bug where
writing to an array element can cause a change to the corresponding
element in a previously made copy of the array.  This bug was
first noticed in Nov 2010 but the fix back then did not go far enough.

bk: 52cb1c7fh_0ZrbdQ4bcdSGNDLv8JUQ